refactoring and sign up

+++ controllers/facebook/facebookConfig.js
@@ -1,21 +1,29 @@
 const { validationResult } = require("express-validator");
 const {  ObjectId } = require('mongodb');
 
-const HttpError = require("../../models/httpError");
 const FacebookConfig = require("../../models/facebook/facebookConfig");
 
 const getFacebookConfig = async (req, res, next) => {
+  const error = validationResult(req);
   const { userId } = req.body;
+  if (!userId) {
+    return res
+    .status(401)
+    .json({ data: {}, message: `Missing userId`, status: false });
+  }
   try {
     const facebookConfigs = await FacebookConfig.find({ userId: userId });
+    const sortedList = facebookConfigs.sort((a, b) => b.createdAt - a.createdAt);
+
     res.json({
-      facebookConfig: facebookConfigs.map((facebookConfig) =>
-        facebookConfig.toObject({ getters: true })
-      ),
+      data: sortedList.map((facebookConfig) =>
+      facebookConfig.toObject({ getters: true })
+    ), status: true, message: 'Fetched data stream'
     });
   } catch (err) {
-    const error = new HttpError("No user found", 500);
-    return next(error);
+    return res
+    .status(500)
+    .json({ data: {}, message: "No config found", status: false });
   }
 };
 
@@ -63,38 +71,51 @@ const saveFacebookConfig = async (req, res, next) => {
       selectedFacebookUser,
       datePreset,
       breakdowns,
-      timeIncrement
+      timeIncrement,
+      createdAt: new Date().getTime(),
     });
 
     await createConfig.save();
   } catch (err) {
-    const error = new HttpError(err, 500);
-    return next(error);
+    return res
+    .status(500)
+    .json({ data: {}, message: "Something went wrong", status: false });
   }
 
-  return res.json({ config: "Config saved successfully" });
+  return res.json({ message: "Config saved successfully", status: true });
 };
 
 const deleteFacebookConfig = async (req, res, next) => {
   const { id } = req.body;
-
+  if (!id) {
+    return res
+    .status(401)
+    .json({ data: {}, message: `Missing id`, status: false });
+  }
   const query = { _id: new ObjectId(id) };
 
   try {
     const result = await FacebookConfig.deleteOne(query);
     if (result.deletedCount === 1) {
-      return res.json({ message: "Document deleted successfully" });
+      return res.json({ message: "Data Stream deleted successfully", status: true });
     } else {
-      return res.json({ message: "Document not found" });
+      return res.json({ message: "Data Stream not found", status: false });
     }
   } catch (err) {
-    const error = new HttpError(err, 500);
-    return next(error);
+    return res
+    .status(500)
+    .json({ data: {}, message: "Something went wrong", status: false });
   }
 };
 
 const updateFacebookConfig = async (req, res, next) => {
   const error = validationResult(req);
+  if (!error.isEmpty()) {
+    return res
+    .status(500)
+    .json({ data: {}, message: `Invalid inputs passed, Please check your data`, status: false });
+  }
+  
   const {
     configName,
     account,
@@ -138,12 +159,17 @@ const updateFacebookConfig = async (req, res, next) => {
       selectedFacebookUser,
       datePreset,
       breakdowns,
-      timeIncrement
+      timeIncrement,
+      createdAt: new Date().getTime(),
     },
   };
 
   const query = { _id: new ObjectId(id) };
-
+  if (!id) {
+    return res
+    .status(401)
+    .json({ data: {}, message: `Missing id`, status: false });
+  }
   try {
     const result = await FacebookConfig.updateOne(query, newData);
     if (result.modifiedCount === 1) {
@@ -152,8 +178,9 @@ const updateFacebookConfig = async (req, res, next) => {
       return res.json({ config: "Document not found or not modified" });
     }
   } catch (err) {
-    const error = new HttpError(err, 500);
-    return next(error);
+    return res
+    .status(500)
+    .json({ data: {}, message: "Something went wrong", status: false });
   }
 };
